feat(binary heap): split the class into two sub classes
Split the binary heap class into two sub classes one for max heap and another one for the min heap
to clean the code and seperate the two behaviors

BREAKING CHANGE: The main class is an abstract class now
